// src/features/lab/types/lmLab.ts
export interface StepDetail {
    step: number;
    char: string;
    probability: number;
}